mapi_lib: avoid mixing signed/unsigned variables around utf8_to_utf16le
HBO on
    f0 (/usr/lib64/libasan.so.4+0x78fb1)
    f1 in EXT_PUSH::p_bytes(void const*, unsigned int) lib/mapi/ext_buffer.cpp:2107
    f2 in EXT_PUSH::p_wstr(char const*) lib/mapi/ext_buffer.cpp:2257
    f3 in p_connect_rsp exch/mh/emsmdb.cpp:890

Fixes: gromox-1.28-76-g9c442510a

+++ include/gromox/util.hpp
@@ -99,7 +99,7 @@ extern void wchar_to_utf8(uint32_t wchar, char *string);
 const char* replace_iconv_charset(const char *charset);
 extern GX_EXPORT BOOL string_to_utf8(const char *charset, const char *in_string, char *out_string, size_t out_len);
 extern BOOL string_from_utf8(const char *charset, const char *in_string, char *out_string, size_t out_len);
-extern int utf8_to_utf16le(const char *src, void *dst, size_t len);
+extern ssize_t utf8_to_utf16le(const char *src, void *dst, size_t len);
 extern BOOL utf16le_to_utf8(const void *src, size_t src_len, char *dst, size_t len);
 extern BOOL get_digest(const char *src, const char *tag, char *buff, size_t buff_len);
 extern BOOL set_digest(char *src, size_t length, const char *tag, const char *value);
